[SPS-174] FIX: 암장 수정 오류 수정 (#113)
* Feat: 난이도 및 암장정보 수정 시 뷰 업데이트

* Feat: attempt patch id 오류 해결

* Fix: 난이도 수정 오류 해결

* Fix: 선택 암장 정보가 남아있는 문제 해결

* [SPS-59] 난이도 및 암장 정보 조회 (#114)

* Fix: cursor 파싱 오류 해결

* Feat: 암장 정보 조회

* Feat: 난이도 조회

+++ Cllog/Projects/CllogService/Data/Data/Sources/DataMapping/Response/BaseResponse/MetaResponseDTO.swift
@@ -8,6 +8,6 @@
 
 import Foundation
 struct BaseMetaResponseDTO: Decodable {
-    let nextCursor: Int?
+    let nextCursor: Double?
     let hasMore: Bool
 }
